pyproject.toml
[project]
name = "mp20"
version = "0.1.0"
description = "MP2.0 planning-first portfolio scaffold"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.1,<6.0",
    "django-cors-headers>=4.4,<5.0",
    "djangorestframework>=3.15,<4.0",
    "pydantic>=2.8,<3.0",
    "psycopg[binary]>=3.2,<4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-django>=4.9,<5.0",
    "ruff>=0.8,<1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "web.mp20_web.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["engine/tests", "web"]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["engine", "extraction", "integrations", "web"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "DJ"]
ignore = ["DJ008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
